philo life

+++ philo_one/srcs/ft_putall.c
@@ -6,7 +6,7 @@
 /*   By: qtamaril <qtamaril@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/09 15:16:57 by qtamaril          #+#    #+#             */
-/*   Updated: 2020/11/09 15:18:12 by qtamaril         ###   ########.fr       */
+/*   Updated: 2020/11/10 11:58:06 by qtamaril         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,18 +38,18 @@ void			ft_putendl_fd(char *s, int fd)
 	ft_putchar_fd('\n', fd);
 }
 
-static void		ft_putnbr_recursion(long int nb, int fd)
+static void		ft_putnbr_recursion(long long nb, int fd)
 {
 	if (nb > 9)
 		ft_putnbr_recursion(nb / 10, fd);
 	ft_putchar_fd(nb % 10 + '0', fd);
 }
 
-void			ft_putnbr_fd(int n, int fd)
+void			ft_putnbr_fd(long int n, int fd)
 {
-	long int	nb;
+	long long	nb;
 
-	nb = (long int)n;
+	nb = (long long)n;
 	if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
